requirements.txt
Quart==0.19.4
Quart-CORS==0.7.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
boto3==1.34.0
azure-mgmt-network==25.1.0
azure-identity==1.15.0
google-cloud-compute==1.16.1
aiohttp==3.9.1
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
hypercorn==0.15.0
marshmallow==3.20.1
aioboto3==12.1.0
cachetools==5.3.2
asyncio-throttle==1.0.2